+++ src/test/kotlin/io/snyk/plugin/ui/toolwindow/SnykToolWindowPanelTest.kt
@@ -15,6 +15,7 @@ import io.snyk.plugin.services.SnykApplicationSettingsStateService
 import io.snyk.plugin.services.SnykTaskQueueService
 import org.eclipse.lsp4j.services.LanguageServer
 import org.eclipse.lsp4j.services.WorkspaceService
+import org.junit.Ignore
 import org.junit.Test
 import snyk.UIComponentFinder
 import snyk.common.lsp.LanguageServerWrapper
@@ -46,6 +47,7 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
         lsw.languageServer = lsMock
         lsw.languageClient = lsClientMock
         lsw.process = lsProcessMock
+        lsw.isInitialized = true
 
         every { lsProcessMock.info().startInstant().isPresent } returns true
         every { lsProcessMock.isAlive } returns true
@@ -151,6 +153,8 @@ class SnykToolWindowPanelTest : LightPlatform4TestCase() {
         verify(exactly = 1) { taskQueueService.scan(false) }
     }
 
+    //TODO rewrite
+    @Ignore("change to language server")
     @Test
     fun `should automatically enable all products on first run after Auth`() {
         val application = ApplicationManager.getApplication()